Move methods from utils.py -> atom_mapping.py
