+++ timemachine/fe/atom_mapping.py
@@ -1,11 +1,12 @@
 from typing import Optional
 
+import networkx as nx
 import numpy as np
 from rdkit import Chem
 from rdkit.Chem import rdFMCS
+from scipy.spatial.distance import cdist
 
 from timemachine.fe.topology import AtomMappingError
-from timemachine.fe.utils import simple_geometry_mapping
 
 
 class CompareDist(rdFMCS.MCSAtomCompare):
@@ -177,4 +178,87 @@ def get_core_by_geometry(mol_a, mol_b, threshold=0.5):
     """
     core = simple_geometry_mapping(mol_a, mol_b, threshold)
     _assert_core_reasonableness(mol_a, mol_b, core)
+    return core
+
+
+def _weighted_adjacency_graph(conf_a, conf_b, threshold=1.0):
+    """construct a networkx graph with
+    nodes for atoms in conf_a, conf_b, and
+    weighted edges connecting (conf_a[i], conf_b[j])
+        if distance(conf_a[i], conf_b[j]) <= threshold,
+        with weight = threshold - distance(conf_a[i], conf_b[j])
+    """
+    distances = cdist(conf_a, conf_b)
+    within_threshold = distances <= threshold
+
+    g = nx.Graph()
+    for i in range(len(within_threshold)):
+        neighbors_of_i = np.where(within_threshold[i])[0]
+        for j in neighbors_of_i:
+            g.add_edge(f"conf_a[{i}]", f"conf_b[{j}]", weight=threshold - distances[i, j])
+    return g
+
+
+def _core_from_matching(matching):
+    """matching is a set of pairs of node names"""
+
+    # 'conf_b[9]' -> 9
+    ind_from_node_name = lambda name: int(name.split("[")[1].split("]")[0])
+
+    match_list = list(matching)
+
+    inds_a = [ind_from_node_name(u) for (u, _) in match_list]
+    inds_b = [ind_from_node_name(v) for (_, v) in match_list]
+
+    return np.array([inds_a, inds_b]).T
+
+
+def core_from_distances(mol_a, mol_b, threshold=1.0):
+    """
+    TODO: docstring
+    TODO: test
+    """
+    # fetch conformer, assumed aligned
+    conf_a = mol_a.GetConformer(0).GetPositions()
+    conf_b = mol_b.GetConformer(0).GetPositions()
+
+    g = _weighted_adjacency_graph(conf_a, conf_b, threshold)
+
+    matching = nx.algorithms.matching.max_weight_matching(g, maxcardinality=True)
+
+    return _core_from_matching(matching)
+
+
+def simple_geometry_mapping(mol_a, mol_b, threshold=0.5):
+    """For each atom i in conf_a, if there is exactly one atom j in conf_b
+    such that distance(i, j) <= threshold, add (i,j) to atom mapping
+
+    Notes
+    -----
+    * Warning! There are many situations where a pair of atoms that shouldn't be mapped together
+        could appear within distance threshold of each other in their respective conformers
+    """
+
+    # fetch conformer, assumed aligned
+    conf_a = mol_a.GetConformer(0).GetPositions()
+    conf_b = mol_b.GetConformer(0).GetPositions()
+    # TODO: perform initial alignment
+
+    within_threshold = cdist(conf_a, conf_b) <= threshold
+    num_neighbors = within_threshold.sum(1)
+    num_mappings_possible = np.prod(num_neighbors[num_neighbors > 0])
+
+    if max(num_neighbors) > 1:
+        print(
+            f"Warning! Multiple (~ {num_mappings_possible}) atom-mappings would be possible at threshold={threshold}Å."
+        )
+        print(f"Only mapping atoms that have exactly one neighbor within {threshold}Å.")
+        # TODO: print more information about difference between size of set returned and set possible
+        # TODO: also assert that only pairs of the same element will be mapped together
+
+    inds = []
+    for i in range(len(conf_a)):
+        if num_neighbors[i] == 1:
+            inds.append((i, np.argmax(within_threshold[i])))
+    core = np.array(inds)
     return core
